% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mqr_methods.R
\name{deviance.mqr}
\alias{deviance.mqr}
\title{Moving quantile regression model deviance}
\usage{
\method{deviance}{mqr}(object, ...)
}
\arguments{
\item{object}{a \code{\link{mqr}} object}

\item{...}{currently nothing}
}
\value{
The sum of squared residuals
}
\description{
Returns the deviance of a fitted moving quantile regression object
}
\author{
Virmantas Kvedaras, Vaidotas Zemlys
}

