% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mqr.R
\name{mqr}
\alias{mqr}
\title{Moving quantile regression}
\usage{
mqr(formula, data, ...)
}
\arguments{
\item{formula}{a formula specifying moving quantile regression}

\item{data}{data data}

\item{...}{additional parameters}
}
\value{
an mqr object
}
\description{
Performs moving quantile regression
}
\details{
Performs moving quantile regression
}
\author{
Vaidotas Zemlys
}

