% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mq.R
\name{mq}
\alias{mq}
\title{Moving quantile for a given band}
\usage{
mq(x, alpha, k, d = NULL, ...)
}
\arguments{
\item{x}{a vector}

\item{alpha}{a quantile}

\item{k}{band structure}

\item{d}{the discount factor}

\item{...}{additional parameters}
}
\value{
a matrix
}
\description{
Calculates a moving quantile for a given band
}
\details{
Given data take a moving band and calculate a quantile based on that band.
}
\author{
Vaidotas Zemlys
}

