% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mqr.R
\docType{package}
\name{mqr}
\alias{mqr}
\alias{mqr-package}
\title{Moving quantile regression}
\usage{
mqr(formula, data, start = NULL, Ofunction = "optim", ...)
}
\arguments{
\item{formula}{a formula specifying moving quantile regression}

\item{data}{data data}

\item{start}{the starting values for discount function}

\item{Ofunction}{the name of optimisation function to use for NLS problem}

\item{...}{additional parameters}
}
\value{
an mqr object
}
\description{
Performs moving quantile regression
}
\details{
Performs moving quantile regression
}
\author{
Vaidotas Zemlys
}

